requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
scipy
google-genai
google-generativeai
python-dotenv